requirements.txt
altgraph==0.17.3
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==3.2.25
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
et-xmlfile==1.1.0
Flask==2.2.3
idna==3.4
importlib-metadata==1.7.0
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
MouseInfo==0.1.3
numpy==1.21.6
oauthlib==3.2.2
openpyxl==3.0.10
pandas==0.25.1
pefile==2023.2.7
Pillow==9.4.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
PyJWT==2.8.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python3-openid==3.2.0
pytweening==1.0.4
pytz==2022.1
pywhatkit==5.4
pywin32==304
pywin32-ctypes==0.2.0
requests==2.28.2
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
soupsieve==2.4
sqlparse==0.4.4
typing-extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.2.3
wikipedia==1.4.0
xlrd==2.0.1
zipp==3.15.0
